Extract the Fixture class.
